* Fix build * Fix comments in README.md * Metadata description can be null * Remove kubernetes overrides in Gopkg.toml
