+++ Gopkg.toml
@@ -38,22 +38,6 @@ required = ["github.com/wadey/gocovmerge"]
   version = "v1.0.5"
 
 [[override]]
-  name = "k8s.io/kubernetes"
-  revision = "v1.8.14"
-
-[[override]]
-  name = "k8s.io/api"
-  revision = "kubernetes-1.8.14"
-
-[[override]]
-  name = "k8s.io/apimachinery"
-  revision = "kubernetes-1.8.14"
-
-[[override]]
-  name = "k8s.io/client-go"
-  revision = "kubernetes-1.8.14"
-
-[[override]]
   name = "github.com/xeipuuv/gojsonschema"
   revision = "93e72a773fade158921402d6a24c819b48aba29d"
