+++ specification/schemas/metadata_schema_v0.2.json
@@ -11,7 +11,10 @@
             "type": "string"
         },
         "description": {
-            "type": "string"
+            "type": [
+                "string",
+                "null"
+            ]
         },
         "namespace": {
             "type": "string"
@@ -44,6 +47,7 @@
         }
     },
     "required": [
-        "name"
+        "name",
+        "version"
     ]
 }
